Moved the SwoUpdater trait to the examples only as it just clutters the main lib with no real benefit
