+++ probe-rs/src/architecture/arm/swo/mod.rs
@@ -1,8 +1,6 @@
 mod decoder;
-mod publisher;
 
 pub use decoder::{Decoder, ExceptionAction, ExceptionType, TracePacket};
-pub use publisher::{SwoPublisher, UpdaterChannel};
 
 use crate::Error;
